application/backend/requirements.txt
Django>=5.0
djangorestframework
celery
django-celery-results
django-cors-headers
gunicorn
psycopg2-binary
